requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.3
defusedxml==0.7.1
distlib==0.3.8
dj-rest-auth==5.0.2
Django==5.0.2
django-allauth==0.61.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-reset-migrations==0.4.0
drf-yasg==1.21.7
filelock==3.13.1
filetype==1.2.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
platformdirs==4.1.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pyotp==2.9.0
python-dateutil==2.8.2
python-decouple==3.8
python-magic==0.4.27
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
sib-api-v3-sdk==7.6.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.0
virtualenvwrapper-win==1.2.7
whitenoise==6.6.0
